+++ src/main/java/pro/zackpollard/telegrambot/api/internal/chat/message/MessageImpl.java
@@ -19,11 +19,11 @@ public class MessageImpl implements Message {
 
     private final int message_id;
     private final User from;
-    private final int date;
+    private final long date;
     private final Chat chat;
     private final User forward_from;
     private final Chat forward_from_chat;
-    private final int forward_date;
+    private final Long forward_date;
     private final Message reply_to_message;
     private final Content content;
 
@@ -37,11 +37,11 @@ public class MessageImpl implements Message {
 
         message_id = jsonObject.getInt("message_id");
         from = UserImpl.createUser(jsonObject.optJSONObject("from"));
-        date = jsonObject.getInt("date");
+        date = jsonObject.getLong("date");
         chat = ChatImpl.createChat(jsonObject.getJSONObject("chat"), telegramBot);
         forward_from = UserImpl.createUser(jsonObject.optJSONObject("forward_from"));
         forward_from_chat = ChatImpl.createChat(jsonObject.optJSONObject("forward_from_chat"), telegramBot);
-        forward_date = jsonObject.optInt("forward_date");
+        forward_date = jsonObject.optLong("forward_date");
         reply_to_message = MessageImpl.createMessage(jsonObject.optJSONObject("reply_to_message"), telegramBot);
         content = ContentImpl.createContent(jsonObject, telegramBot);
 
@@ -69,7 +69,7 @@ public class MessageImpl implements Message {
      * @return The timestamp
      */
     @Override
-    public int getTimeStamp() {
+    public long getTimeStamp() {
         return date;
     }
 
@@ -119,7 +119,7 @@ public class MessageImpl implements Message {
      * @return The forwarded messages timestamp, or 0 if the message wasn't forwarded
      */
     @Override
-    public int getForwardedDate() {
+    public Long getForwardedDate() {
         return forward_date;
     }
